requirements.txt
apiview==1.3.25
asgiref==3.2.7
astroid==2.4.1
Babel==2.8.0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==6.7
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
DateTime==4.3
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.6
django-better-admin-arrayfield==1.4.0
django-colorfield==0.3.0
django-cors-headers==3.2.1
django-environ==0.4.5
django-import-export==2.4.0
django-map-widgets==0.3.0
django-phonenumber-field==4.0.0
django-rest-swagger==2.2.0
django-storages==1.9.1
django-twilio==0.13.0.post1
djangorestframework==3.11.0
docopt==0.6.2
drf-yasg==1.17.1
et-xmlfile==1.0.1
feedparser==5.2.1
google-api-core==1.17.0
google-auth==1.14.2
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
gunicorn==20.0.4
idna==2.9
inflection==0.4.0
isort==4.3.21
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.2
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.21
numpy==1.19.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.5
packaging==20.3
parso==0.7.1
phonenumbers==8.12.2
Pillow==7.2.0
pluggy==0.13.1
ply==3.11
prompt-toolkit==2.0.10
protobuf==3.11.3
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.1
PyJWT==1.7.1
pyparsing==2.4.7
python-dotenv==0.14.0
python-http-client==3.3.1
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.5.2
sendgrid==6.4.7
Shapely==1.7.1
simplejson==3.17.2
six==1.14.0
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
tablib==2.0.0
twilio==6.39.0
ujson==3.2.0
uritemplate==3.0.1
uritemplate.py==3.0.2
urllib3==1.25.9
wcwidth==0.2.5
whitenoise==5.2.0
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
django-heroku==0.3.1
django-better-admin-arrayfield==1.4.0
django-map-widgets==0.3.0
zope.interface==5.1.2